// src/model/user.ts
// "_id": "5cb05b4db6ed8c44f42c9af2",
//         "username": "test",
//         "password": "202cb962ac59075b964b07152d234b70",
//         "phone": "123412342134",
//         "email": "sd",
//         "role_id": "5ca9eab0b49ef916541160d4",
//         "create_time": 1555061581734,

export interface User {
  _id: string;
  username: string;
  password: string;
  phone: string;
  email: string;
  role_id: string;
  create_time: number;
}